refactor(lib): move url
