+++ app.js
@@ -29,7 +29,7 @@ const controllers = require('./app/controllers')
 const requestHandlers = require('./lib/request_handlers')
 const initCloudinary = require('./lib/cloudinary')
 const compileSVGSprites = require('./lib/svg-sprite')
-const appURL = require('./lib/url')
+const appURL = require('./app/lib/url')
 const apiRoutes = require('./app/api_routes')
 const serviceRoutes = require('./app/service_routes')
 const logger = require('./lib/logger.js')
