+++ lib/request_handlers/metatags.js
@@ -1,7 +1,7 @@
 const request = require('request')
 const { User, Street } = require('../../app/db/models')
 const logger = require('../../lib/logger.js')
-const appURL = require('../../lib/url.js')
+const appURL = require('../../app/lib/url')
 
 const ANON_CREATOR = '-'
